Split sigstore configuration parsing and API into separate files
... to make policy_config*.go a bit smaller, and to allow sigstore
logic expansion.

Only moves unchanged code, should not change behavior.
